Remove "dangerous" OS X specific .gitignore rule
The .gitignore rule "Icon?" will also match directories such as
"icons/", at least on a case-insensitive file-system (this happened to
me on OS X 10.7). I only noticed this several days later because a
developer was complaining about missing icons in his repository. I don't
really know what type of files this rule is supposed to match, but I
guess the described behavior is not expected.