Slashify Ruby files and directories.
Perhaps this is a little pedantic, but without the initial / all
such matches within the repository get ignored, and without the
trailing slash files as well as directories of that name get ignored.
e.g. Specifying `tmp` and `test/tmp` without slashes is redundant, as
`tmp` already ignores `test/tmp`, as well as every other `tmp` and
`tmp/`.