Update JetBrains.gitignore
In newer versions of the IntelliJ platform (such as the one used for Rider), the .idea files are now placed in a subdirectory of the same, named after the current project (such as ".idea/.idea.Everlook/.idea/workspace.xml"). Adding a double-star pattern to these folder rules will also ignore these new files and maintain backwards compatibility with previous revisions of this file.