Add Cursive's REPL history file
**Reasons for making this change:**

The file `replstate.xml` contains the history of the Clojure REPL
that Cursive adds to IntelliJ. Obviously that's user-specific,
and not relevant to other users.

**Links to documentation supporting these rule changes:**

This file is not well-documented, but in cursive-ide/cursive#1325,
the Cursive developers state that this is the REPL history file,
and that deleting it is acceptable troubleshooting if it's
causing trouble.